+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unreleased
+
+### Bugfix
+
+- Prevent edge case where the player's faction would not be removed from the available enemy faction pool
+
 ## v6.5.0 - 2026-07-30
 
 ### Added